+++ dynamic/react/api-app/user-manager.js
