+++ dynamic/react/shared/components/apiConsole.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ConsoleInputForm from './consoleInputForm';
 import ConsoleLiveData from './consoleLiveData';
-import userManager from '../user-manager';
 
 // Helper that determines what part of the endpoint is shown in the `Request` input of
 // the ConsoleLiveData component
@@ -18,18 +17,6 @@ const getRequest = (endpoint) => {
 const ApiConsole = (props) => {
     return (
         <div>
-            {props.userProfile ?
-                <button className={'btn btn-secondary'} onClick={() => {
-                    sessionStorage.devdotRedirectUrl = window.location.href;
-                    userManager.signoutRedirect();
-                }} style={{margin: '8px'}}>{'Logout'}</button> :
-                <button className={'btn btn-primary'} onClick={() => {
-                    sessionStorage.devdotRedirectUrl = window.location.href;
-                    userManager.signinRedirect();
-                }} style={{margin: '8px'}}>{'Authorize'}
-                </button>
-            }
-
             <div className={'row api-console'}>
                 <div className={'col-md-4 col-xs-12 api-console-form-wrapper'}>
                     <ConsoleInputForm {...props} />
